pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "photoholmes"
version = "0.0.1"
requires-python = ">=3.10"
license = { file = "LICENSE" }
dependencies = [
    "numpy",
    "matplotlib",
    "opencv-python",
    "scikit-learn",
    "scikit-image",
    "pydantic",
    "tqdm",
    "jpegio",
    "scipy",
    "pyyaml",
    "mpmath",
    "typer",
    "pillow",
    "wget",
    "ipykernel",

    "torch >= 2.1",
    "torchmetrics",
    "torchvision",
    "torch_kmeans",
]

[project.optional-dependencies]
dev = ["pytest", "black", "isort", "pre-commit", "flake8", "mypy"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["photoholmes*"]

[project.scripts]
photoholmes = "photoholmes.cli.main:cli"

[tool.black]
target-version = ['py310']

[tool.isort]
profile = 'black'

# ===== MYPY configuration ====== #
[tool.mypy]
ignore_missing_imports = true
show_error_codes = true


[[tool.mypy.overrides]]
module = ["torch.*"]
follow_imports = "skip"
follow_imports_for_stubs = true
# =============================== #
